requirements.txt
PyYAML
pymysql